Updated some of the code for UserRole functionality. Long typed id instead of Integer/int

+++ PROJECT-HUB/projecttwobase/src/main/java/com/example/projecttwobase/repository/UserRoleRepository.java
@@ -3,6 +3,6 @@ package com.example.projecttwobase.repository;
 import com.example.projecttwobase.model.UserRole;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRoleRepository extends CrudRepository<UserRole, Integer>  {
+public interface UserRoleRepository extends CrudRepository<UserRole, Long>  {
     UserRole findByName(String name);
 }
